Firebase Work done

+++ app/src/main/java/com/example/eduthon_aimers/Donation/Donor.kt
@@ -4,7 +4,7 @@ package com.example.eduthon_aimers.Donation
 data class Donor(
     var DonorName:String,
     var DonorEmailAddress:String,
-    var DonorPhoneNumber:Int,
+    var DonorPhoneNumber:Long,
     var DonorAddress:String,
     var DonorCity:String,
     var DonorState:String,
