+++ app/src/main/java/com/example/eduthon_aimers/Donation/Seeker.kt
@@ -3,7 +3,7 @@ package com.example.eduthon_aimers.Donation
 class Seeker (
     val seekerName:String,
     var seekerEmail:String,
-    var seekerPhone:Int,
+    var seekerPhone:Long,
     var seekerPinCode:Int,
     var seekerCity:String,
     var seekerState:String,
